Frap & HA Overhaul Part 1 + Tweaks to screens and logic
FRAP and HA has been updated to fit the revisions(90% complete)
Notifications has returned for now, which will now display ADmin notifs as well albeit a bit weird.
Deductions Page buttons cannot be pressed until 10th or 25th of the days.
+forgotten bug fixes ill remember later

Things left to do: Combine the logics of the Docs, Doc Versions and the applications.